network, tests: specify test lanes != k8s.1{6|7} as dual-stack
Also remove the sriov lane from the dual stack lanes.

All the most representative of a tests will be tested in dual
stack mode.

For instance, follow-up commits in this PR focus on the service
tests, and will create - and check connectivity towards - an ipv6
service.

To execute the tests manually against an external cluster, in dual
stack mode, you should add the `-dual-stack` flag when calling the
tests binary:

```
${TESTS_OUT_DIR}/tests.test \
    -kubeconfig=${kubeconfig} \
...
    -dual-stack
```

Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>